Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-2
> 
> While this might be appropriate, it has not major effect.  The vendor blobs portion is missing.  The 15.1 missed that as well.

You're referring to it not being committed to the favored storage area?  This was partially meant as a hint to whomever knew where the controls for that were to go ahead and take care of this issue.

The major point was to cleanup every other build since this pollutes 80% of the builds using lge_msm8996-common which don't require these files.

Patch-set: 2
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>